Refactor nazw
