+++ integration-tests/src/test/java/com/pip/chatbot/integration/tests/AdminAuthorizationTest.java
@@ -1,8 +1,8 @@
-package com.pip.chatbot.controller;
+package com.pip.chatbot.integration.tests;
 
 import com.intuit.karate.junit5.Karate;
 
-public class AdminControllerTest {
+public class AdminAuthorizationTest {
 
     @Karate.Test
     Karate test(){
