--- integration-tests/src/test/java/com/pip/chatbot/integration/tests/SampleTest.java
@@ -1,10 +1,0 @@
-package com.pip.chatbot.integration.tests;
-import com.intuit.karate.junit5.Karate;
-
-class SampleTest {
-    @Karate.Test
-    Karate testSample() {
-        return Karate.run().relativeTo(getClass());
-    }
-}
-
