--- integration-tests/src/test/java/com/pip/chatbot/integration/tests/sample.feature
@@ -1,7 +1,0 @@
-Feature: sample
-
-  Scenario: first hello world
-    * print 'hello'
-
-  Scenario: second scenario
-    * print 'world'
