+++ integration-tests/src/test/java/karate-config.js
@@ -2,7 +2,7 @@ function fn() {
     return {
         user: 'admin',
         password: 'Admin123*pip-2020',
-        uncorrectPassword: 'Admin123*pip+2020',
+        incorrectPassword: 'Admin123*pip+2020',
         baseUrl: "http://localhost:9090"
     }
 }
